Avance equipment entity

+++ ournms/Model/Equipment.cs
@@ -1,10 +1,10 @@
 namespace ournms.Model;
 
 
-public class Equipment(int id, string name, string ipAddress)
+public class Equipment(long id, string name, string ipAddress)
 {
 
-    public int Id { get; set; } = id;
+    public long Id { get; set; } = id;
     public string Name { get; set; } = name;
     public string IpAddress { get; set; } = ipAddress;
 }
